+++ java/client/src/main/java/glide/api/BaseClient.java
@@ -3,8 +3,6 @@ package glide.api;
 
 import static glide.api.models.GlideString.gs;
 import static glide.api.models.commands.SortBaseOptions.STORE_COMMAND_STRING;
-import static glide.api.models.commands.bitmap.BitFieldOptions.BitFieldReadOnlySubCommands;
-import static glide.api.models.commands.bitmap.BitFieldOptions.BitFieldSubCommands;
 import static glide.api.models.commands.bitmap.BitFieldOptions.createBitFieldArgs;
 import static glide.api.models.commands.bitmap.BitFieldOptions.createBitFieldGlideStringArgs;
 import static glide.api.models.commands.stream.StreamClaimOptions.JUST_ID_REDIS_API;
@@ -148,6 +146,8 @@ import static redis_request.RedisRequestOuterClass.RequestType.XGroupCreateConsu
 import static redis_request.RedisRequestOuterClass.RequestType.XGroupDelConsumer;
 import static redis_request.RedisRequestOuterClass.RequestType.XGroupDestroy;
 import static redis_request.RedisRequestOuterClass.RequestType.XGroupSetId;
+import static redis_request.RedisRequestOuterClass.RequestType.XInfoConsumers;
+import static redis_request.RedisRequestOuterClass.RequestType.XInfoGroups;
 import static redis_request.RedisRequestOuterClass.RequestType.XLen;
 import static redis_request.RedisRequestOuterClass.RequestType.XPending;
 import static redis_request.RedisRequestOuterClass.RequestType.XRange;
@@ -1763,7 +1763,7 @@ public abstract class BaseClient
 
     @Override
     public CompletableFuture<Object> invokeScript(@NonNull Script script) {
-        if (script.getBinarySafeOutput()) {
+        if (script.getBinaryOutput()) {
             return commandManager.submitScript(
                     script, List.of(), List.of(), this::handleBinaryObjectOrNullResponse);
         } else {
@@ -1775,7 +1775,7 @@ public abstract class BaseClient
     @Override
     public CompletableFuture<Object> invokeScript(
             @NonNull Script script, @NonNull ScriptOptions options) {
-        if (script.getBinarySafeOutput()) {
+        if (script.getBinaryOutput()) {
             return commandManager.submitScript(
                     script,
                     options.getKeys().stream().map(GlideString::gs).collect(Collectors.toList()),
@@ -1793,7 +1793,7 @@ public abstract class BaseClient
     @Override
     public CompletableFuture<Object> invokeScript(
             @NonNull Script script, @NonNull ScriptOptionsGlideString options) {
-        if (script.getBinarySafeOutput()) {
+        if (script.getBinaryOutput()) {
             return commandManager.submitScript(
                     script, options.getKeys(), options.getArgs(), this::handleBinaryObjectOrNullResponse);
         } else {
@@ -2707,6 +2707,40 @@ public abstract class BaseClient
                         new GlideString[] {gs(JUST_ID_REDIS_API)});
         return commandManager.submitNewCommand(
                 XClaim, args, response -> castArray(handleArrayResponse(response), GlideString.class));
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>[]> xinfoGroups(@NonNull String key) {
+        return commandManager.submitNewCommand(
+                XInfoGroups,
+                new String[] {key},
+                response -> castArray(handleArrayResponse(response), Map.class));
+    }
+
+    @Override
+    public CompletableFuture<Map<GlideString, Object>[]> xinfoGroups(@NonNull GlideString key) {
+        return commandManager.submitNewCommand(
+                XInfoGroups,
+                new GlideString[] {key},
+                response -> castArray(handleArrayResponseBinary(response), Map.class));
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>[]> xinfoConsumers(
+            @NonNull String key, @NonNull String groupName) {
+        return commandManager.submitNewCommand(
+                XInfoConsumers,
+                new String[] {key, groupName},
+                response -> castArray(handleArrayResponse(response), Map.class));
+    }
+
+    @Override
+    public CompletableFuture<Map<GlideString, Object>[]> xinfoConsumers(
+            @NonNull GlideString key, @NonNull GlideString groupName) {
+        return commandManager.submitNewCommand(
+                XInfoConsumers,
+                new GlideString[] {key, groupName},
+                response -> castArray(handleArrayResponseBinary(response), Map.class));
     }
 
     @Override