+++ python/python/tests/test_async_client.py
@@ -76,7 +76,8 @@ from glide.config import (
     ProtocolVersion,
     ServerCredentials,
 )
-from glide.constants import OK, TEncodable, TResult
+from glide.constants import OK, TEncodable, TFunctionStatsResponse, TResult
+from glide.exceptions import TimeoutError as GlideTimeoutError
 from glide.glide_client import GlideClient, GlideClusterClient, TGlideClient
 from glide.routes import (
     AllNodes,
@@ -91,10 +92,12 @@ from glide.routes import (
 from tests.conftest import create_client
 from tests.utils.utils import (
     check_function_list_response,
+    check_function_stats_response,
     check_if_server_version_lt,
     compare_maps,
     convert_bytes_to_string_object,
     convert_string_to_bytes_object,
+    create_lua_lib_with_long_running_function,
     generate_lua_lib_code,
     get_first_result,
     get_random_string,
@@ -8086,6 +8089,272 @@ class TestCommands:
         with pytest.raises(RequestError) as e:
             await glide_client.function_delete(lib_name)
         assert "Library not found" in str(e)
+
+    @pytest.mark.parametrize("cluster_mode", [False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_function_stats(self, glide_client: GlideClient):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(glide_client, min_version):
+            return pytest.mark.skip(reason=f"Redis version required >= {min_version}")
+
+        lib_name = "functionStats"
+        func_name = lib_name
+        assert await glide_client.function_flush(FlushMode.SYNC) == OK
+
+        # function $funcName returns first argument
+        code = generate_lua_lib_code(lib_name, {func_name: "return args[1]"}, False)
+        assert await glide_client.function_load(code, True) == lib_name.encode()
+
+        response = await glide_client.function_stats()
+        check_function_stats_response(response, [], 1, 1)
+
+        code = generate_lua_lib_code(
+            lib_name + "_2",
+            {func_name + "_2": "return 'OK'", func_name + "_3": "return 42"},
+            False,
+        )
+        assert (
+            await glide_client.function_load(code, True) == (lib_name + "_2").encode()
+        )
+
+        response = await glide_client.function_stats()
+        check_function_stats_response(response, [], 2, 3)
+
+        assert await glide_client.function_flush(FlushMode.SYNC) == OK
+
+        response = await glide_client.function_stats()
+        check_function_stats_response(response, [], 0, 0)
+
+    @pytest.mark.parametrize("cluster_mode", [True])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_function_stats_cluster(self, glide_client: GlideClusterClient):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(glide_client, min_version):
+            return pytest.mark.skip(reason=f"Redis version required >= {min_version}")
+
+        lib_name = "functionStats_without_route"
+        func_name = lib_name
+        assert await glide_client.function_flush(FlushMode.SYNC) == OK
+
+        # function $funcName returns first argument
+        code = generate_lua_lib_code(lib_name, {func_name: "return args[1]"}, False)
+        assert await glide_client.function_load(code, True) == lib_name.encode()
+
+        response = await glide_client.function_stats()
+        for node_response in response.values():
+            check_function_stats_response(
+                cast(TFunctionStatsResponse, node_response), [], 1, 1
+            )
+
+        code = generate_lua_lib_code(
+            lib_name + "_2",
+            {func_name + "_2": "return 'OK'", func_name + "_3": "return 42"},
+            False,
+        )
+        assert (
+            await glide_client.function_load(code, True) == (lib_name + "_2").encode()
+        )
+
+        response = await glide_client.function_stats()
+        for node_response in response.values():
+            check_function_stats_response(
+                cast(TFunctionStatsResponse, node_response), [], 2, 3
+            )
+
+        assert await glide_client.function_flush(FlushMode.SYNC) == OK
+
+        response = await glide_client.function_stats()
+        for node_response in response.values():
+            check_function_stats_response(
+                cast(TFunctionStatsResponse, node_response), [], 0, 0
+            )
+
+    @pytest.mark.parametrize("cluster_mode", [True])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    @pytest.mark.parametrize("single_route", [True, False])
+    async def test_function_stats_with_routing(
+        self, glide_client: GlideClusterClient, single_route: bool
+    ):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(glide_client, min_version):
+            return pytest.mark.skip(reason=f"Redis version required >= {min_version}")
+
+        route = (
+            SlotKeyRoute(SlotType.PRIMARY, get_random_string(10))
+            if single_route
+            else AllPrimaries()
+        )
+        lib_name = "functionStats_with_route_" + str(single_route)
+        func_name = lib_name
+        assert await glide_client.function_flush(FlushMode.SYNC, route) == OK
+
+        # function $funcName returns first argument
+        code = generate_lua_lib_code(lib_name, {func_name: "return args[1]"}, False)
+        assert await glide_client.function_load(code, True, route) == lib_name.encode()
+
+        response = await glide_client.function_stats(route)
+        if single_route:
+            check_function_stats_response(
+                cast(TFunctionStatsResponse, response), [], 1, 1
+            )
+        else:
+            for node_response in response.values():
+                check_function_stats_response(
+                    cast(TFunctionStatsResponse, node_response), [], 1, 1
+                )
+
+        code = generate_lua_lib_code(
+            lib_name + "_2",
+            {func_name + "_2": "return 'OK'", func_name + "_3": "return 42"},
+            False,
+        )
+        assert (
+            await glide_client.function_load(code, True, route)
+            == (lib_name + "_2").encode()
+        )
+
+        response = await glide_client.function_stats(route)
+        if single_route:
+            check_function_stats_response(
+                cast(TFunctionStatsResponse, response), [], 2, 3
+            )
+        else:
+            for node_response in response.values():
+                check_function_stats_response(
+                    cast(TFunctionStatsResponse, node_response), [], 2, 3
+                )
+
+        assert await glide_client.function_flush(FlushMode.SYNC, route) == OK
+
+        response = await glide_client.function_stats(route)
+        if single_route:
+            check_function_stats_response(
+                cast(TFunctionStatsResponse, response), [], 0, 0
+            )
+        else:
+            for node_response in response.values():
+                check_function_stats_response(
+                    cast(TFunctionStatsResponse, node_response), [], 0, 0
+                )
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_function_kill_no_write(
+        self, request, cluster_mode, protocol, glide_client: TGlideClient
+    ):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(glide_client, min_version):
+            return pytest.mark.skip(reason=f"Redis version required >= {min_version}")
+
+        lib_name = f"mylib1C{get_random_string(5)}"
+        func_name = f"myfunc1c{get_random_string(5)}"
+        code = create_lua_lib_with_long_running_function(lib_name, func_name, 10, True)
+
+        # nothing to kill
+        with pytest.raises(RequestError) as e:
+            await glide_client.function_kill()
+        assert "NotBusy" in str(e)
+
+        # load the library
+        assert await glide_client.function_load(code, replace=True) == lib_name.encode()
+
+        # create a second client to run fcall
+        test_client = await create_client(
+            request, cluster_mode=cluster_mode, protocol=protocol, timeout=15000
+        )
+
+        # call fcall to run the function
+        # make sure that fcall routes to a primary node, and not a replica
+        # if this happens, function_kill and function_stats won't find the function and will fail
+        primaryRoute = SlotKeyRoute(SlotType.PRIMARY, lib_name)
+
+        async def endless_fcall_route_call():
+            # fcall is supposed to be killed, and will return a RequestError
+            with pytest.raises(RequestError) as e:
+                if cluster_mode:
+                    await test_client.fcall_ro_route(
+                        func_name, arguments=[], route=primaryRoute
+                    )
+                else:
+                    await test_client.fcall_ro(func_name, arguments=[])
+            assert "Script killed by user" in str(e)
+
+        async def wait_and_function_kill():
+            # it can take a few seconds for FCALL to register as running
+            await asyncio.sleep(3)
+            timeout = 0
+            while timeout <= 5:
+                # keep trying to kill until we get an "OK"
+                try:
+                    result = await glide_client.function_kill()
+                    #  we expect to get success
+                    assert result == "OK"
+                    break
+                except RequestError:
+                    # a RequestError may occur if the function is not yet running
+                    # sleep and try again
+                    timeout += 0.5
+                    await asyncio.sleep(0.5)
+
+        await asyncio.gather(
+            endless_fcall_route_call(),
+            wait_and_function_kill(),
+        )
+
+        # no functions running so we get notbusy error again
+        with pytest.raises(RequestError) as e:
+            assert await glide_client.function_kill()
+        assert "NotBusy" in str(e)
+
+    @pytest.mark.parametrize("cluster_mode", [False, True])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_function_kill_write_is_unkillable(
+        self, request, cluster_mode, protocol, glide_client: TGlideClient
+    ):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(glide_client, min_version):
+            return pytest.mark.skip(reason=f"Redis version required >= {min_version}")
+
+        lib_name = f"mylib1C{get_random_string(5)}"
+        func_name = f"myfunc1c{get_random_string(5)}"
+        code = create_lua_lib_with_long_running_function(lib_name, func_name, 10, False)
+
+        # load the library on all primaries
+        assert await glide_client.function_load(code, replace=True) == lib_name.encode()
+
+        # create a second client to run fcall - and give it a long timeout
+        test_client = await create_client(
+            request, cluster_mode=cluster_mode, protocol=protocol, timeout=15000
+        )
+
+        # call fcall to run the function loaded function
+        async def endless_fcall_route_call():
+            # fcall won't be killed, because kill only works against fcalls that don't make a write operation
+            # use fcall(key) so that it makes a write operation
+            await test_client.fcall(func_name, keys=[lib_name])
+
+        async def wait_and_function_kill():
+            # it can take a few seconds for FCALL to register as running
+            await asyncio.sleep(3)
+            timeout = 0
+            foundUnkillable = False
+            while timeout <= 5:
+                # keep trying to kill until we get a unkillable return error
+                try:
+                    await glide_client.function_kill()
+                except RequestError as e:
+                    if "UNKILLABLE" in str(e):
+                        foundUnkillable = True
+                        break
+                timeout += 0.5
+                await asyncio.sleep(0.5)
+            # expect an unkillable error
+            assert foundUnkillable
+
+        await asyncio.gather(
+            endless_fcall_route_call(),
+            wait_and_function_kill(),
+        )
 
     @pytest.mark.parametrize("cluster_mode", [True])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])